requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.1.0
Flask-Mail==0.9.1
python-dotenv==0.19.0
PyMySQL==1.0.2